+++ color/index.js
@@ -1,7 +1,9 @@
+import getColor from '../getColor'
 import toHex from '../toHex'
 import toRGB from '../toRGB'
 
 const DOM = {
+  getColor,
   toHex,
   toRGB
 }