+++ tests/style/setStyles.spec.js
@@ -5,6 +5,7 @@ import setStyles from '@/setStyles'
 import getStyles from '@/getStyles'
 import byId from '@/byId'
 import toRGB from '@/toRGB'
+import getColor from '../../getColor'
 
 describe('setStyles() 方法', () => {
   // Set up our document body
@@ -38,14 +39,18 @@ describe('setStyles() 方法', () => {
     expect(setStyles()).toBe(false)
   })
 
-  it(`setStyles($list, 'background-color:#fff;color:#000;font-size:13px;') styles 参数采用 cssText 字符串一次设置多个样式：`, () => {
-    setStyles($list, 'background-color:#fff;color:#000;font-size:13px;')
+  it(`setStyles($list, 'background-color:#fff;font-size:13px;') styles 参数采用 cssText 字符串一次设置多个样式：`, () => {
+    setStyles($list, 'background-color:#fff;font-size:13px;')
 
     const rules = getStyles($list, ['background-color', 'color', 'font-size'])
 
-    expect(toRGB(rules.backgroundColor)).toEqual('rgb(255, 255, 255)')
-    expect(toRGB(rules.color)).toEqual('rgb(0, 0, 0)')
+    expect(getColor($list, 'background-color', false)).toEqual('rgb(255, 255, 255)')
+    expect(getColor($list, 'color')).toEqual('transparent')
     expect(rules.fontSize).toEqual('13px')
+
+    setStyles($list, 'color:black;')
+    expect(getColor($list, 'color')).toEqual('#000000')
+    expect(getColor($list, ['color'])).toBe(false)
   })
 
   it(`setStyles($list, {'background-color':'#fff',color:'black','font-size': '13px'}) styles 参数采用 Object 对象格式设置多个样式：`, () => {